+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### 🎁 New Features
 
+* Hoist's enhanced autosizing is now enabled on all grids by default.  See `GridModel`
+and `GridAutosizeService` for more details.
+
 * `DimensionChooser` now has the ability to persist its value and history seperately.
 * Mobile `select` now supports `enableFilter` and `enableCreate`. 
 
@@ -12,8 +15,12 @@
 
 * Fixes to Context Menu handling on `Panel` to allow better handling of `[]` and `null`.
 
-### ⚙️ Technical
-* Additional updates to experimental autosize feature.
+
+### ⚙️ Breaking Changes
+* `emptyFlexCol` has been removed from the Hoist API and should simply be removed from all
+client applications. Improvements to agGrid's default rendering of empty space have made it
+obsolete.
+
 
 [Commit Log](https://github.com/xh/hoist-react/compare/v33.3.0...develop)